requirements.txt
streamlit
sentence-transformers
torch
numpy
faiss-cpu
requests